Km double형으로 바꾸기

+++ app/src/main/java/com/example/survirun/UserModel.java
@@ -7,5 +7,5 @@ public class UserModel {
 
     public int todayCalorie;
     public int todayExerciseTime;
-    public int todayKm;
+    public double todayKm;
 }
